feat(dom): add CMakeList and change directory structure
